+++ libs/app/help/feature/src/help.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { RecommendRouting } from './recommend.routing';
+import { RecommendRouting } from './help.routing';
 import { IonicModule } from '@ionic/angular';
 import { NzStepsModule } from 'ng-zorro-antd/steps';
 import { NzListModule } from 'ng-zorro-antd/list';
@@ -9,7 +9,7 @@ import { NzFormModule } from 'ng-zorro-antd/form';
 import { NzInputModule } from 'ng-zorro-antd/input';
 import { NzIconModule } from 'ng-zorro-antd/icon';
 
-import { RecommendPage } from './recommend.page';
+import { RecommendPage } from './help.page';
 import { NavigationBarModule } from '@fridge-to-plate/app/navigation/feature';
 import { RecommendDataAccessModule } from '@fridge-to-plate/app/recommend/data-access';
 import { RecommendUIModule } from '@fridge-to-plate/app/recommend/ui';
