+++ libs/app/help/feature/src/help.page.html
@@ -1,3 +1,3 @@
 <div class="w-full h-full bg-background-light overflow-y-auto">
-  <stepper-form class="lg:mt-16 lg:mb-0 sm:mb-20 md:mb-20"/>
+
 </div>
