+++ libs/app/help/feature/src/index.ts
@@ -1,1 +1,1 @@
-export * from './recommend.module';
+export * from './help.module';
